fix(schranken): drei Pruefungen, die der Typ schon entschieden hatte (MF-988)
Der Linux-Bau meldet 22-mal `-Wtype-limits` — „comparison is always
true/false due to limited range of data type". Der Uebersetzer sagt damit,
dass **der Typ die Pruefung entscheidet, nicht der Wert**.

Harmlos ist das, solange der Typ ENGER ist als die Schranke: die Pruefung
ist dann ueberfluessig. Gefaehrlich ist die Gegenrichtung — **eine
Schranke, die der Typ nicht erreichen kann, ist keine Schranke.** Sie
sieht im Quelltext aus wie eine Sicherung und ist keine.

Alle 22 einzeln nachgesehen, nicht hochgerechnet:

  6  echte Verhaltensfehler   PETSCII, `uft_file_ops_extended.c`
  1  entwerteter Waechter     JVC,     `uft_jvc_plugin.c`
  1  entwerteter Waechter     SCL,     `uft_scl_parser_v2.c`
  5  32-Bit-Ueberlaufwaechter (`n > SIZE_MAX/sizeof(T)`), auf 64 Bit
     inaktiv und RICHTIG — sie feuern auf 32-Bit-Zielen. Bleiben.
  9  Schranke weiter als der Typ (`uint8_t > 1024`), folgenlos. Bleiben.

**1) JVC — der Cast schnitt ab, BEVOR die Schranke prueft.**

    uint32_t total_tracks = body / track_bytes;
    uint8_t cyls = (uint8_t)(total_tracks / sides);
    if (cyls == 0 || cyls > UFT_MAX_CYLINDERS) { abweisen }

`UFT_MAX_CYLINDERS` ist 256, `cyls` ist ein `uint8_t` — die zweite
Bedingung konnte nie zutreffen. Sie haette auch nichts genuetzt: der Cast
eine Zeile davor hatte den zu grossen Wert bereits in einen kleinen,
plausiblen verwandelt. Die Schranke prueft das Ergebnis der Verstuemmelung.

Gemessen an einer Datei mit 300 Spuren: `UFT_OK` und **44 Zylinder** —
eine voellig gewoehnliche Zahl fuer eine TRS-80-Diskette, 256 Spuren still
verloren. `total_sectors` liegt damit um Faktor 7 daneben. **JVC ist kein
toter Code:** `uft_format_plugin_jvc` steht in der Registry (Z. 282).

Jetzt wird der GELESENE Wert geprueft und erst danach verengt; `cylinders`
im Plugin-Zustand ist `uint16_t`. Abgewiesen wird trotzdem, statt 300 zu
melden — UFT_MAX_CYLINDERS ist die vereinbarte Obergrenze des Baums, und
eine Absage ist ehrlich, eine gekuerzte Zahl nicht.

**2) PETSCII — `char` ist auf x86-64 vorzeichenbehaftet.**

    char name[17];
    if (name[i] == 0xA0) ...                       // nie wahr
    else if (name[i] >= 0xC1 && name[i] <= 0xDA)   // nie wahr / immer wahr

Ein `char` deckt -128..127 ab; 0xA0 ist 160, 0xC1 ist 193. **Die
PETSCII-Wandlung lief durch, ohne ein einziges Zeichen anzufassen** — an
zwei Stellen: Diskettenname und Dateinamen im Verzeichnis.

Das Besondere daran: mit vorzeichenlosem `char` (Linux/ARM) wandelt
dieselbe Quelle korrekt. **Ein Diskettenname sah je nach Bauplattform
anders aus** — in einem forensischen Werkzeug eine stille Veraenderung,
und eine, die kein Test dieses Baums bemerkt haette.

Reichweite ehrlich: `uft_list_files_extended()` hat heute keinen Aufrufer.
`d81_list_files()` ist aber in `include/uft/uft_file_ops.h` oeffentlich
erklaert; der Test ruft die veroeffentlichte API, keinen Umweg.

**3) SCL — ohne Rotbeweis behoben, und das steht so im Code.**

    uint8_t file_count;  #define SCL_MAX_FILES 256
    if (builder->file_count >= SCL_MAX_FILES) return false;

`>= 256` kann fuer einen `uint8_t` nie zutreffen. Beim 257. Eintrag laeuft
der Zaehler auf 0 zurueck und ueberschreibt still den ersten — kein
Speicherfehler (das Feld hat genau 256 Plaetze), aber stiller Datenverlust
mit Erfolgsmeldung. `scl_builder_add_file` ist `static` und hat im ganzen
Baum **keinen Aufrufer**; es gibt kein Verhalten, das rot werden koennte.
Belegt durch Rechnung, so berichtet, Feld auf `uint16_t`.

**Abnahme** — `tests/test_schranken_die_nie_greifen.c`, fuenf Faelle,
davon zwei Gegenproben. Vor den Fixes drei rot, mit Klartext:

    meldet 44 Zylinder fuer eine Datei mit 300 (256 Spuren still verloren)
    Name kam als D5 C6 D4 A0 A0 A0 zurueck, erwartet "UFT"

Mutationsmatrix — jede faellt GENAU ihren Fall, beide Gegenproben bleiben
durchgehend gruen:

  M1  JVC-Cast zurueck        -> nur „meldet nie eine ungelesene Zylinderzahl"
  M2  Diskettenname auf char  -> nur „petscii_wird_in_ascii_gewandelt"
  M3  Dateiname auf char      -> nur „dateiname_wird_aus_petscii_gewandelt"

**Der fuenfte Fall verdankt sich der Matrix.** Mit vier Faellen fiel bei M3
nichts um — der Fix an der zweiten PETSCII-Stelle waere ohne Rotbeweis
geblieben. Die Matrix hat die Luecke gezeigt, nicht der Plan.

Die Gegenproben sind nicht Zierde: ohne „gewoehnliche JVC-Datei meldet 40"
waere „lehne alles ab" ein gruener Fix; ohne „ASCII-Name bleibt
unveraendert" waere „ziehe von allem 0x80 ab" einer. Der ASCII-Fall fuellt
mit 0x20 und nicht mit 0xA0 — mit 0xA0 haenge er an derselben Ursache wie
der Hauptfall und traenne nichts mehr. Auch das kam aus dem ersten Lauf.

Kennzahl (Regel 9): keine der vier. Bugfixes an Bestehendem, EINFRIER-REGEL
nicht beruehrt. Prinzipienfall — „Keine stille Veraenderung".

+++ src/formats/scl/uft_scl_parser_v2.c
@@ -255,7 +255,16 @@ typedef struct {
     uint8_t* data;
     size_t capacity;
     size_t size;
-    uint8_t file_count;
+    /* MF-988: `uint8_t` konnte SCL_MAX_FILES (256) nie erreichen — die
+     * Schranke in scl_builder_add_file() war damit wirkungslos, und beim
+     * 257. Eintrag lief der Zaehler auf 0 zurueck und ueberschrieb still
+     * den ersten. Kein Speicherfehler (das Feld hat genau 256 Plaetze),
+     * aber ein stiller Datenverlust mit Erfolgsmeldung.
+     *
+     * Ohne Rotbeweis behoben, und das steht hier so: `scl_builder_add_file`
+     * ist `static` und hat im ganzen Baum keinen Aufrufer — es gibt kein
+     * Verhalten, das rot werden koennte. Belegt durch Rechnung. */
+    uint16_t file_count;
 } scl_builder_t;
 
 /**
